Renamed system to world.
Previously, code and documentation referred to the ECS data structure as
a system, but also the system functions as systems, which was confusing.
To alleviate this problem, the ECS data structure has been renamed to
"world" in code and documentation. This is a reference to other
similar functional systems in which state is simulated by worlds and
fucntions which take worlds and return worlds, like Racket's Universe
Package, and a Clojure project I can't remmeber the name of.