Changed the treatment of dosync in dynamic mode.
Previously, functions operating in dynamic mode had to handle their own
dosync, and dynamic mode involved a lot of entering and exiting dosync
blocks. This is theoretically fast, because the ref is never shared
cross-thread, but nevertheless, it's a bad idea. So, dynamic mode is now
assumed to be inside a dosync block. This adds the additional feature
that attempting to call dynamic mode functions while not in dynamic mode
will cause them to throw a very clear error message about the use of
alter outside a dosync block, rather than the nil errors that were
previously thrown when the dynamic vars were found to be incorrectly
bound (as is the default case when not in dynamic mode.)